requirements.txt
# Core dependencies
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
requests==2.31.0
python-dotenv==1.0.0

# Web Dashboard
flask==3.0.0
flask-cors==4.0.0

# Optional: For testing
pytest==7.4.3
pytest-asyncio==0.21.1